// src/service-api/generated/models/dealsDealStageDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * sales-crm
 * OpenAPI spec version: 1.0.0
 */

export type DealsDealStageDto = typeof DealsDealStageDto[keyof typeof DealsDealStageDto];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DealsDealStageDto = {
  NEW: 'NEW',
  QUALIFIED: 'QUALIFIED',
  PROPOSAL: 'PROPOSAL',
  WON: 'WON',
  LOST: 'LOST',
} as const;