// src/service-api/generated/models/dealsListItemDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * sales-crm
 * OpenAPI spec version: 1.0.0
 */
import type { DealsListItemDtoNextTask } from './dealsListItemDtoNextTask';
import type { DealsListItemDtoContact } from './dealsListItemDtoContact';
import type { DealsListItemCompanyDto } from './dealsListItemCompanyDto';
import type { DealsDealStageDto } from './dealsDealStageDto';

export interface DealsListItemDto {
  /** @nullable */
  nextTask: DealsListItemDtoNextTask;
  /** @nullable */
  contact: DealsListItemDtoContact;
  company: DealsListItemCompanyDto;
  /** @nullable */
  closeDate: string | null;
  stage: DealsDealStageDto;
  value: string;
  title: string;
  id: string;
}